update: 调整 type 目录结构
